Fixed animation of removal of matched cards
First, I put removeMatchedCards in the abstract view controller. This is because the set view controller uses too much of it.
Then, I called it from handleSpecificTapInController

In addition, for the animation to successeed, the overall view must not be updated to be without the matched cards until the animation finishes